Component refactoring + Codacy + PropTypes
