+++ src/app/component/CartItem/CartItem.container.js
@@ -14,9 +14,9 @@ import { connect } from 'react-redux';
 import React, { PureComponent } from 'react';
 import { ProductType } from 'Type/ProductList';
 import { CartDispatcher } from 'Store/Cart';
+import { convertKeyValueObjectToQueryString } from 'Util/Url';
 
 import CartItem from './CartItem.component';
-import { convertKeyValueObjectToQueryString } from 'Util/Url';
 
 export const mapDispatchToProps = dispatch => ({
     addProduct: options => CartDispatcher.addProductToCart(dispatch, options),
@@ -31,15 +31,35 @@ export class CartItemContainer extends PureComponent {
 
         this.containerFunctions = {
             handleQtyChange: this.handleQtyChange.bind(this),
-            handleRemoveItem: this.handleRemoveItem.bind(this),
-            getProductLinkTo: this._getProductLinkTo.bind(this),
-            getProductThumbnail: this._getProductThumbnail.bind(this)
+            handleRemoveItem: this.handleRemoveItem.bind(this)
         };
 
         this.containerProps = () => ({
             thumbnail: this._getProductThumbnail(),
             linkTo: this._getProductLinkTo()
         });
+    }
+
+    handleQtyChange(value) {
+        const { addProduct, product, product: { quantity } } = this.props;
+        const newQuantity = value - quantity;
+
+        if (newQuantity) {
+            this.setState({ isLoading: true });
+            addProduct({ product, quantity: newQuantity }).then(
+                () => this.setState({ isLoading: false })
+            );
+        }
+    }
+
+    handleRemoveItem() {
+        const { removeProduct, product } = this.props;
+
+        this.setState({ isLoading: true });
+
+        removeProduct({ product }).then(
+            () => this.setState({ isLoading: false })
+        );
     }
 
     /**
@@ -87,29 +107,6 @@ export class CartItemContainer extends PureComponent {
 
         return thumbnail ? `/media/catalog/product${ thumbnail }` : '';
     }
-
-    handleQtyChange(value) {
-        const { addProduct, product, product: { quantity } } = this.props;
-        const newQuantity = value - quantity;
-
-        if (newQuantity) {
-            this.setState({ isLoading: true });
-            addProduct({ product, quantity: newQuantity }).then(
-                () => this.setState({ isLoading: false })
-            );
-        }
-    }
-
-    handleRemoveItem() {
-        const { removeProduct, product } = this.props;
-
-        this.setState({ isLoading: true });
-
-        removeProduct({ product }).then(
-            () => this.setState({ isLoading: false })
-        );
-    }
-
 
     render() {
         return (
