+++ src/app/component/Html/index.js
@@ -9,4 +9,4 @@
  * @link https://github.com/scandipwa/base-theme
  */
 
-export { default } from './Html.component';
+export { default } from './Html.container';
