+++ src/app/component/ProductPrice/ProductPrice.container.js
@@ -10,8 +10,8 @@
  */
 
 import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
 import { PriceType } from 'Type/ProductList';
+import { MixType } from 'Type/Common';
 import ProductPrice from './ProductPrice.component';
 /**
  * Product price
@@ -70,14 +70,7 @@ class ProductPriceContainer extends PureComponent {
 
 ProductPriceContainer.propTypes = {
     price: PriceType,
-    mix: PropTypes.shape({
-        block: PropTypes.string,
-        elem: PropTypes.string,
-        mods: PropTypes.objectOf(PropTypes.oneOfType([
-            PropTypes.string,
-            PropTypes.bool
-        ]))
-    })
+    mix: MixType
 };
 
 ProductPriceContainer.defaultProps = {
